Fix huffman decoding with table 11 (ESC). We were storing the quantized coefficients in int8_t, but that's not wide enough when there are escapes.

+++ AacSpectrumDecoder.cpp
@@ -100,7 +100,6 @@ bool AacSpectrumDecoder::decode2(unsigned int tableNum, int out[2])
 
   while (true)
   {
-    //printf("AacSpectrumDecoder::decode2(): i %d  len %d  codeword 0x%X  v1 %d  v2 %d\n", i, len, codeword, huffmanTable->entries[i].v0, huffmanTable->entries[i].v1);
     // If we've hit an entry with more bits than we have, read more bits
     if (len < huffmanTable->entries[i].len)
     {
@@ -113,10 +112,11 @@ bool AacSpectrumDecoder::decode2(unsigned int tableNum, int out[2])
     // Check each entry of the current length
     while (huffmanTable->entries[i].len == len)
     {
+      //printf("AacSpectrumDecoder::decode2(): i %d  len %d  codeword 0x%X  v1 %d  v2 %d\n", i, len, codeword, huffmanTable->entries[i].v0, huffmanTable->entries[i].v1);
       if (huffmanTable->entries[i].codeword == codeword)
       {
-        int8_t v0 = huffmanTable->entries[i].v0;
-        int8_t v1 = huffmanTable->entries[i].v1;
+        int v0 = huffmanTable->entries[i].v0;
+        int v1 = huffmanTable->entries[i].v1;
 
         // Read sign bits if needed, but don't act on them yet
         bool sign0 = false;
@@ -170,7 +170,6 @@ bool AacSpectrumDecoder::decode4(unsigned int tableNum, int out[4])
 
   while (true)
   {
-    //printf("AacSpectrumDecoder::decode4(): i %d  len %d  codeword 0x%X  v1 %d  v2 %d  v3 %d  v4 %d\n", i, len, codeword, huffmanTable->entries[i].v0, huffmanTable->entries[i].v1, huffmanTable->entries[i].v2, huffmanTable->entries[i].v3);
     // If we've hit an entry with more bits than we have, read more bits
     if (len < huffmanTable->entries[i].len)
     {
@@ -183,12 +182,13 @@ bool AacSpectrumDecoder::decode4(unsigned int tableNum, int out[4])
     // Check each entry of the current length
     while (huffmanTable->entries[i].len == len)
     {
+      //printf("AacSpectrumDecoder::decode4(): i %d  len %d  codeword 0x%X  v1 %d  v2 %d  v3 %d  v4 %d\n", i, len, codeword, huffmanTable->entries[i].v0, huffmanTable->entries[i].v1, huffmanTable->entries[i].v2, huffmanTable->entries[i].v3);
       if (huffmanTable->entries[i].codeword == codeword)
       {
-        int8_t v0 = huffmanTable->entries[i].v0;
-        int8_t v1 = huffmanTable->entries[i].v1;
-        int8_t v2 = huffmanTable->entries[i].v2;
-        int8_t v3 = huffmanTable->entries[i].v3;
+        int v0 = huffmanTable->entries[i].v0;
+        int v1 = huffmanTable->entries[i].v1;
+        int v2 = huffmanTable->entries[i].v2;
+        int v3 = huffmanTable->entries[i].v3;
 
         if (!codebooks[tableNum].isSigned)
         {
